Fix incorrect insets padding
