+++ app/src/main/java/io/nekohasekai/sagernet/tasker/TaskerActivity.kt
@@ -74,17 +74,6 @@ class TaskerActivity : ThemedActivity(R.layout.layout_config_settings),
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val toolbar = findViewById<Toolbar>(R.id.toolbar)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
         setSupportActionBar(toolbar)
         supportActionBar?.apply {
             setTitle(R.string.tasker_settings)
