+++ app/src/main/java/io/nekohasekai/sagernet/ui/AboutFragment.kt
@@ -48,17 +48,6 @@ class AboutFragment : ToolbarFragment(R.layout.layout_about) {
         binding = LayoutAboutBinding.bind(view)
 
         toolbar.setTitle(R.string.menu_about)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
         ViewCompat.setOnApplyWindowInsetsListener(binding.layoutAbout) { v, insets ->
             val bars = insets.getInsets(
                 WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
