+++ app/src/main/java/io/nekohasekai/sagernet/ui/AssetEditActivity.kt
@@ -104,18 +104,6 @@ class AssetEditActivity(
         super.onCreate(savedInstanceState)
 
         val toolbar: Toolbar = findViewById(R.id.toolbar)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
-
         setSupportActionBar(toolbar)
         supportActionBar?.apply {
             setTitle(R.string.assets_settings)
