+++ app/src/main/java/io/nekohasekai/sagernet/ui/AssetsActivity.kt
@@ -65,18 +65,6 @@ class AssetsActivity : ThemedActivity(), UndoSnackbarManager.Interface<File> {
         updating = findViewById(R.id.action_updating)
         updating.isGone = true
 
-        val toolbar = findViewById<Toolbar>(R.id.toolbar)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
         ViewCompat.setOnApplyWindowInsetsListener(binding.recyclerView) { v, insets ->
             val bars = insets.getInsets(
                 WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
@@ -88,6 +76,8 @@ class AssetsActivity : ThemedActivity(), UndoSnackbarManager.Interface<File> {
             )
             insets
         }
+
+        val toolbar = findViewById<Toolbar>(R.id.toolbar)
         setSupportActionBar(toolbar)
         supportActionBar?.apply {
             setTitle(R.string.route_assets)
