+++ app/src/main/java/io/nekohasekai/sagernet/ui/GroupSettingsActivity.kt
@@ -231,18 +231,6 @@ class GroupSettingsActivity(
         super.onCreate(savedInstanceState)
 
         val toolbar = findViewById<Toolbar>(R.id.toolbar)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
-
         setSupportActionBar(toolbar)
         supportActionBar?.apply {
             setTitle(R.string.group_settings)
