+++ app/src/main/java/io/nekohasekai/sagernet/ui/SettingsFragment.kt
@@ -13,17 +13,6 @@ class SettingsFragment : ToolbarFragment(R.layout.layout_config_settings) {
         super.onViewCreated(view, savedInstanceState)
 
         toolbar.setTitle(R.string.settings)
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
 
         if (savedInstanceState == null) parentFragmentManager.beginTransaction()
             .replace(R.id.settings, SettingsPreferenceFragment())
