+++ app/src/main/java/io/nekohasekai/sagernet/ui/configuration/ConfigurationFragment.kt
@@ -197,19 +197,6 @@ class ConfigurationFragment @JvmOverloads constructor(
                 requireActivity().finish()
             }
         }
-
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val bars = insets.getInsets(
-                WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
-            )
-            v.updatePadding(
-                top = bars.top,
-                left = bars.left,
-                right = bars.right,
-            )
-            insets
-        }
-
         val activity = activity
         val searchView = toolbar.findViewById<SearchView>(R.id.action_search)
         if (searchView != null) {
