+++ app/src/main/java/io/nekohasekai/sagernet/ui/dashboard/ConnectionFragment.kt
@@ -56,11 +56,6 @@ class ConnectionFragment() :
                 (requireActivity() as MainActivity).onBackPressedCallback.isEnabled = true
             }
         }
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, insets ->
-            val statusBarTop = insets.getInsets(WindowInsetsCompat.Type.statusBars()).top
-            v.updatePadding(top = statusBarTop)
-            insets
-        }
         ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
             val bars = insets.getInsets(
                 WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout()
